requirements.txt
streamlit>=1.36
polygon-api-client>=1.16.1
yfinance>=0.2.40
pandas>=2.2.0
matplotlib>=3.8.0
python-dateutil>=2.9.0.post0
exchange-calendars>=4.11.2
